Update sources in response to systemd-networkd events (LP: #1718227).
This is reusing the already provided script for NetworkManager.
This script is not hard-tied to NetworkManager as it does not use
and of its provided environment variables.

OTOH the NM scripts are general scripts that are called on any change
and then differentiate between invocation reasons via the environment provided.
But the script for chrony is agnostic to the event types - it just checks if
servers are reachable and calls chronyc online/offline accordingly.

Therefore we just link this script into networkd-dispatcher events to work the
same way. Even if not setting servers offline via loosing an interface, but
instead e.g. using the auto_offline function of chrony and then configuring a
new device it will detect the servers as no-reachable and try to connect the
sources again.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>