- removed d/init to avoid weird interactions between sysV and systemd
We already fixed the "remove chrony" case, but there are more.
Install-Remove timesyncd - ok
Install chrony - ok
Install-Purge-Install chrony - ok
Install-Remove-Install-Reinstall chrony - ok
Install-Remove-Install - timesyncd ??

The reason for the latter is that out of Config-File state [1] it
gets called with a version number. That makes it use
  invoke-rc.d chrony restart
instead of
  invoke-rc.d chrony start
as it does on the fresh install.

But that restart is much softer, it considers the service to be masked
due to the dh_installsystemd section comes later.

In fact if you'd switch the order of dh_installsystemd and dh_installinit
it would work but it would cause other side effects.

We had similar (not the same, but also weird sysV/systemd interactions)
problems in other packages and the only real way out is to do the cut
and move over to systemd for real.

Therefore dropping the sysV init script for chrony

[1]: https://wiki.debian.org/MaintainerScripts

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>