debian/tests/upstream-simulation-test-suite: redirect stderr on make
On some architectures the clksim tests compile but throw some warnings.
E.g. on armhf
 2452 |       _Rb_tree<_Key, _Val, _KeyOfValue, _Compare, _Alloc>::
      |       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /usr/include/c++/9/map:61,
                 from generator.h:23,
                 from clock.h:23,
                 from node.h:22,
                 from network.h:26,
                 from network.cc:19:

Now it is fine to abort if the make fails (bad RC + set -e), but if it
is just a warning let the test continue. Therefore redirect the stderr
of the make call to stdout.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>