rework NM-dispatcher/dhclient detection
The packaged NM dispatcher and dhclient scripts provide the same
functionality. Only one should be active depending on the system
configuration.

Instead of guessing in the NM dispatcher script whether the dhclient
script is executed, check in the dhclient script whether it is called
from a NM dispatcher script (the 11-dhclient script included in the
dhcp-client package).

This is simpler and it removes dependency on the 11-dhclient script.