Drop obsolete workaround in scriptlet
The bug with daemon-reload was fixed a long time ago.
The direct invocation of 'systemctl daemon-reload' with no redirect of
errors to /dev/null causes a user-visible message when installing into a chroot:

  Running scriptlet: chrony-4.2-4.fc37.x86_64     320/334
System has not been booted with systemd as init system (PID 1). Can't operate.
Failed to connect to bus: Host is down
  ...